Move some variables out of model
There were some stuff that's not necessary to be in model
and can just remain unexported variables.

```release-note
NONE
```
