+++ model/post.go
@@ -69,8 +69,6 @@ const (
 	PostPropsGroupHighlightDisabled   = "disable_group_highlight"
 )
 
-var AtMentionPattern = regexp.MustCompile(`\B@`)
-
 type Post struct {
 	Id         string `json:"id"`
 	CreateAt   int64  `json:"create_at"`
